remove pcap dep on drawin and windows OS
