+++ agent/sysinfo/pcap_darwin.go
@@ -1,0 +1,8 @@
+package sysinfo
+
+import "github.com/lodastack/agent/agent/common"
+
+// PcapMetrics dep pcap lib
+func PcapMetrics() (L []*common.Metric) {
+	return
+}
